requirements.txt
flask
tensorflow-cpu
numpy
pillow
gunicorn
werkzeug
pytest
